Move structs scattered in `hash_engine_lib` to `experiment` module
