+++ packages/engine/lib/orchestrator/src/process/local.rs
@@ -8,8 +8,7 @@ use async_trait::async_trait;
 use error_stack::{IntoReport, Report, ResultExt};
 use execution::package::experiment::ExperimentId;
 use hash_engine_lib::{
-    experiment::controller::run::cleanup_experiment,
-    proto::EngineMsg,
+    experiment::{comms::EngineMsg, controller::run::cleanup_experiment},
     utils::{LogFormat, LogLevel, OutputLocation},
 };
