+++ packages/engine/src/experiment/comms/mod.rs
@@ -1,0 +1,7 @@
+mod engine;
+mod orchestrator;
+
+pub use self::{
+    engine::{EngineMsg, InitMessage},
+    orchestrator::{OrchClient, OrchestratorMsg},
+};
