+++ packages/engine/src/experiment/controller/mod.rs
@@ -1,11 +1,8 @@
 pub mod config;
 pub mod controller;
-mod error;
 pub mod run;
 pub mod sim_configurer;
 
 // TODO: Currently unused, left in in case Rust runner will use it. Can safely be removed
 //   afterwards.
 // pub mod utils;
-
-pub use self::error::{Error, Result};
