+++ packages/engine/src/experiment/error.rs
@@ -1,17 +1,36 @@
+use error_stack::Report;
 use thiserror::Error as ThisError;
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
 
 #[derive(ThisError, Debug)]
 pub enum Error {
-    #[error("{0}")]
+    #[error("Experiment Controller error: {0}")]
     Unique(String),
+
+    #[error("Stateful error: {0}")]
+    Stateful(#[from] stateful::Error),
 
     #[error("Execution error: {0}")]
     Execution(#[from] execution::Error),
 
-    #[error("Controller error: {0}")]
-    Controller(#[from] crate::experiment::controller::Error),
+    #[error("Structure error: {0}")]
+    Structure(#[from] experiment_structure::Error),
+
+    #[error("Simulation error: {0}")]
+    SimulationControl(#[from] simulation_control::Error),
+
+    #[error("Simulation controller error: {0}")]
+    SimulationController(#[from] simulation_control::controller::Error),
+
+    #[error("Serialize/Deserialize error")]
+    Serde(#[from] serde_json::Error),
+
+    #[error("Nano error: {0:?}")]
+    Nano(Report<nano::ErrorKind>),
+
+    #[error("Missing configuration in dynamic payloads. Key: {0}")]
+    MissingConfiguration(String),
 
     #[error("Simulation run's changed Global values are not in a JSON object")]
     ChangedGlobalsNotObject,
@@ -24,6 +43,9 @@ pub enum Error {
 
     #[error("Property is not object, but is supposed to contain a varying property: {0}")]
     NestedPropertyNotObject(String),
+
+    #[error("Unexpected message to the engine, expected an init message")]
+    UnexpectedEngineMsgExpectedInit,
 }
 
 impl From<&str> for Error {
@@ -35,5 +57,11 @@ impl From<&str> for Error {
 impl From<String> for Error {
     fn from(s: String) -> Self {
         Error::Unique(s)
+    }
+}
+
+impl From<Report<nano::ErrorKind>> for Error {
+    fn from(report: Report<nano::ErrorKind>) -> Self {
+        Self::Nano(report)
     }
 }
