+++ packages/engine/src/lib.rs
@@ -35,7 +35,6 @@
 )]
 
 mod args;
-pub mod env;
 pub mod experiment;
 pub mod proto;
 pub mod utils;
