+++ packages/engine/src/proto.rs
@@ -1,21 +1,13 @@
 use execution::{
-    package::{experiment::ExperimentId, simulation::SimulationId},
+    package::simulation::SimulationId,
     runner::{
         comms::{PackageError, UserError, UserWarning},
         RunnerError,
     },
 };
-use experiment_structure::ExperimentRun;
 use serde::{Deserialize, Serialize};
 use simulation_control::SimStatus;
 use stateful::global::Globals;
-
-/// The message type sent from the engine to the orchestrator.
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OrchestratorMsg {
-    pub experiment_id: ExperimentId,
-    pub body: EngineStatus,
-}
 
 #[derive(Serialize, Deserialize, Debug, PartialEq)]
 pub enum EngineStatus {
@@ -36,41 +28,6 @@ pub enum EngineStatus {
     UserWarnings(SimulationId, Vec<UserWarning>),
     PackageError(SimulationId, PackageError),
     Logs(SimulationId, Vec<String>),
-}
-
-/// The message type sent from the orchestrator to the engine.
-#[derive(Serialize, Deserialize, Debug)]
-pub enum EngineMsg {
-    Init(InitMessage),
-}
-
-/// The initialization message sent by an Orchestrator implementation to the Engine
-#[derive(Serialize, Deserialize, Debug)]
-pub struct InitMessage {
-    /// Defines the type of Experiment that's being ran (e.g. a wrapper around a single-run of a
-    /// simulation, or the configuration for a normal experiment)
-    pub experiment: ExperimentRun,
-    /// Unused
-    pub env: ExecutionEnvironment,
-    /// A JSON object of dynamic configurations for things like packages, see
-    /// [`OUTPUT_PERSISTENCE_KEY`] for an example
-    ///
-    /// [`OUTPUT_PERSISTENCE_KEY`]: crate::experiment::controller::[`OUTPUT_PERSISTENCE_KEY`]
-    pub dyn_payloads: serde_json::Map<String, serde_json::Value>,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub enum ExecutionEnvironment {
-    Local { port: u16 },
-    Staging,
-    Production,
-    None,
-}
-
-impl Default for ExecutionEnvironment {
-    fn default() -> Self {
-        ExecutionEnvironment::None
-    }
 }
 
 impl EngineStatus {
